♻️ refactor : update,delete,detail view을 합쳐 url 하나로 연결
